Updated todo file
Improved the speed of the local pirate bay searching to the point where multi threading factories are no longer needed

Improved the sorting algorithm to prioritise seeders over things like ratio so popular things rise to the top

Added progress dialog to easytv so user can't do anything until the client has connected to the server and fixed the close code so when an error is shown on startup dismissing it actually closes down the application

Added stop download and pause api methods

Added serializable fields to torrent download and fixed infinite GSON recursion bug

Made use of Debugger.log instead of Message.message for the category parameter.

Slightly altered S/L ratio calculation

+++ Client/src/com/mt/easytv/torrents/SearchScore.java
@@ -7,11 +7,11 @@ public class SearchScore
     private int    relevanceScore;
     private int    seedersScore;
     private int    ratioScore;
-    private double overallScore;
+    private int overallScore;
     private double relevance;
     private double ratio;
 
-    public SearchScore(int relevanceScore, int seedersScore, int ratioScore, double overallScore, double relevance, double ratio) {
+    public SearchScore(int relevanceScore, int seedersScore, int ratioScore, int overallScore, double relevance, double ratio) {
         this.relevanceScore = relevanceScore;
         this.seedersScore = seedersScore;
         this.ratioScore = ratioScore;
@@ -22,7 +22,7 @@ public class SearchScore
 
     public static SearchScore fromMap(Map map) {
         return new SearchScore(((Number) map.get("relevanceScore")).intValue(), ((Number) map.get("seedersScore")).intValue(),
-                               ((Number) map.get("ratioScore")).intValue(), ((Number) map.get("overallScore")).doubleValue(),
+                               ((Number) map.get("ratioScore")).intValue(), ((Number) map.get("overallScore")).intValue(),
                                ((Number) map.get("relevance")).doubleValue(), ((Number) map.get("ratio")).doubleValue());
     }
